hw 18 blogs sa fix1

+++ src/features/sa/use-cases/deleteSaBlogs.ts
@@ -1,9 +1,12 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { BlogsSaService } from '../api/blogs.sa.service';
-import { NotFoundException } from '@nestjs/common';
+import { ForbiddenException, NotFoundException } from '@nestjs/common';
 
 export class DeleteSaBlogs {
-  constructor(public readonly blogId: string) {}
+  constructor(
+    public readonly blogId: string,
+    public readonly userId: string,
+  ) {}
 }
 
 @CommandHandler(DeleteSaBlogs)
@@ -12,6 +15,8 @@ export class DeleteSaBlogsHandler implements ICommandHandler<DeleteSaBlogs> {
   async execute(command: DeleteSaBlogs): Promise<any> {
     const findBlog = await this.blogsSaService.findBlogId(command.blogId);
     if (!findBlog) throw new NotFoundException();
+
+    if (findBlog.userId !== command.userId) throw new ForbiddenException();
     return this.blogsSaService.deleteBlogId(command.blogId);
   }
 }
